Fix UBSan signed-int overflow in flameGraph dealloc for size = INT64_MIN
In `AggregateFunctionFlameGraph.cpp:418`, the dealloc branch computed
`UInt64 abs_size = -size;` which is undefined behaviour when
`size == INT64_MIN` because the magnitude `2^63` is not representable
in `Int64`.

Cast to `UInt64` first so the negation runs in unsigned arithmetic,
where it is well-defined two's complement and yields the correct
absolute value. Matches the sibling pattern on the allocation path
(`allocation->size = UInt64(size);`).

Report: https://s3.amazonaws.com/clickhouse-test-reports/json.html?REF=master&sha=544f6da6dc6cc737e7f8fc8e3bcf7e7d5657f099&name_0=MasterCI&name_1=Stress%20test%20%28experimental%2C%20serverfuzz%2C%20arm_asan_ubsan%29
STID: 3765-3f7f

+++ src/AggregateFunctions/AggregateFunctionFlameGraph.cpp
@@ -415,7 +415,9 @@ struct AggregateFunctionFlameGraphData
         }
         else if (size < 0)
         {
-            UInt64 abs_size = -size;
+            /// Cast to unsigned before negating: -size is UB when size == INT64_MIN
+            /// because INT64_MAX + 1 is not representable in Int64.
+            UInt64 abs_size = -static_cast<UInt64>(size);
             if (auto * allocation = tryFindMatchAndRemove(place.allocation, abs_size))
             {
                 untrack(allocation);
